+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## Unreleased
+- Added UUIDs for Lessons and Grades
+- Added a v2 API for Lessons and Grades that uses UUIDs for identifiers
 - Upgraded Ruby runtime to 2.6.0
 - Enable showing deleted students in single group view